Refactored names to cahnge Report with Dmer, and renamed DMER package path to dmer
