+++ DMFT/src/main/java/com/gov/rsi/DMFT/controllers/AbstractController.java
@@ -1,4 +1,4 @@
-package com.gov.rsi.DMFT.controllers;
+package com.gov.rsi.dmft.controllers;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
