+++ DMFT/src/test/java/com/gov/rsi/DMFT/DmftApplicationTests.java
@@ -1,4 +1,4 @@
-package com.gov.rsi.DMFT;
+package com.gov.rsi.dmft;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
